Changed API to use MethodView subclasses
