+++ app/__init__.py
@@ -2,6 +2,8 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_mail import Mail
 from config import config
+from .api import api as api_blueprint
+from .api.resources import PostView, UserView
 
 
 db = SQLAlchemy()
@@ -14,7 +16,13 @@ def create_app(config_name):
 	db.init_app(app)
 	mail.init_app(app)
 
-	from .api import api as api_blueprint
+	posts_as_view = PostView.as_view('posts')
+	users_as_view = UserView.as_view('users')
+
+	api_blueprint.add_url_rule('/posts', view_func=posts_as_view, methods=['GET', 'POST'])
+	api_blueprint.add_url_rule('/posts/<int:_id>', view_func=posts_as_view, methods=['GET', 'PUT', 'DELETE'])
+	api_blueprint.add_url_rule('/users', view_func=users_as_view, methods=['POST'])
+	api_blueprint.add_url_rule('/confirm/<token>', view_func=users_as_view, methods=['GET'])
 
 	app.register_blueprint(api_blueprint, url_prefix='/api/v1')
