+++ app/api/__init__.py
@@ -1,6 +1,3 @@
 from flask import Blueprint
-from ..models import User
 
 api = Blueprint('api', __name__)
-
-from . import users, posts
